Just moving stuff around and removing Resolution (Umbraco internal) from the Merchello core.
