+++ src/Merchello.sln
@@ -36,6 +36,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Merchello.Web.UI", "Merchel
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Merchello.Plugin.Shipping.FixedRate", "Merchello.Plugin.Shipping.FixedRate\Merchello.Plugin.Shipping.FixedRate.csproj", "{DC644111-E837-43C1-B84C-1E1B27A6DAF7}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Plugins", "Plugins", "{D3BA503A-71E6-4F96-BC79-69BD2A342400}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shipping", "Shipping", "{D2C2BDB2-29A6-41BF-BF0B-7F38BDD64CF3}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -142,5 +146,7 @@ Global
 		{AC1A12FD-34CD-488A-90A1-528DE5941C33} = {A6C6EBCD-38C9-4FAD-975A-97FA97578537}
 		{4A7FCB29-A39A-4E0B-B0BC-92A7D00FA3AB} = {A6C6EBCD-38C9-4FAD-975A-97FA97578537}
 		{7EFA61BD-198A-4CD9-8176-C266FD1D4E1D} = {A6C6EBCD-38C9-4FAD-975A-97FA97578537}
+		{D2C2BDB2-29A6-41BF-BF0B-7F38BDD64CF3} = {D3BA503A-71E6-4F96-BC79-69BD2A342400}
+		{DC644111-E837-43C1-B84C-1E1B27A6DAF7} = {D2C2BDB2-29A6-41BF-BF0B-7F38BDD64CF3}
 	EndGlobalSection
 EndGlobal
