+++ test/Merchello.Tests.Base/Merchello.Tests.Base.csproj
@@ -61,6 +61,7 @@
     <Compile Include="Respositories\UnitOfWork\MockDatabaseUnitOfWork.cs" />
     <Compile Include="Respositories\UnitOfWork\TransactionType.cs" />
     <Compile Include="Services\ServiceTestsBase.cs" />
+    <Compile Include="SqlSyntax\Resolution.cs" />
     <Compile Include="TypeFields\TypeFieldMock.cs" />
     <Compile Include="SqlSyntax\BaseUsingSqlServerSyntax.cs" />
     <Compile Include="Umbraco\ApplicationContext\FakeUmbracoApplication.cs" />
